backend/pyproject.toml
[project]
name = "tabsensei"
version = "3.0.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.120.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.0",
]
